Implement more enum features
